[BUG] Inconsistent HTTP status code on query mismatch
The logical behavour of a router should return http status code of 404 when a request fails to stisfy a route validation logics. Before this, mux was returning 405 http status code in some rare scenarios which is not a valid on its case.

For more info, See: https://github.com/gorilla/mux/issues/704