Use "redirect" rather than "re-direct" in docs (#735)
This brings these docs in line with the rest of the repository.

## What type of PR is this? (check all applicable)

- [ ] Refactor
- [ ] Feature
- [ ] Bug Fix
- [ ] Optimization
- [x] Documentation Update
- [ ] Go Version Update
- [ ] Dependency Update

## Description

Remove use of hyphen in "redirect" in the documentation.

## Added/updated tests?

- [ ] Yes
- [x] No, and this is why: No regression test required 
- [ ] I need help with writing tests

## Run verifications and test

- [ ] `make verify` is passing
- [x] `make test` is passing

`make verify` is not passing for me locally, but the failures are in
unrelated
files. It may be because I don't see a .golangci.yml in the repo. I
assume it's
using my linting config instead.